fetch-ai/requirements.txt
uagents>=0.22.0
uagents-core>=0.2.0
openai>=1.0.0
pydantic>=2.5.0
python-dotenv>=1.2.0
requests>=2.32.0
elasticsearch>=8.0.0
